refactor(render): split trooper renderer module
